+++ core/CHANGELOG.md
@@ -13,8 +13,8 @@
 ### Added
 - [#733](https://github.com/embedded-graphics/embedded-graphics/pull/733) Add envelope() method to `Rectangle`
 - [#735](https://github.com/embedded-graphics/embedded-graphics/pull/735) Implement `Dimensions` for Slices, Arrays, and Tuples
-- [#735](https://github.com/embedded-graphics/embedded-graphics/pull/735) Implement `PointsIter` for Arrays, Tuples, immutable references, and mutable references
-- [#735](https://github.com/embedded-graphics/embedded-graphics/pull/735) Implement `Drawable` for Arrays, Tuples, immutable references, and mutable references
+- [#735](https://github.com/embedded-graphics/embedded-graphics/pull/735) Implement `PointsIter` for Arrays, and Tuples
+- [#735](https://github.com/embedded-graphics/embedded-graphics/pull/735) Implement `Drawable` for Arrays, and Tuples
 
 ## [0.4.0] - 2023-05-14